通用Generator

+++ zheng-pay/zheng-pay-dao/src/main/java/com/zheng/pay/dao/mapper/PayInOrderDetailMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface PayInOrderDetailMapper {
-    int countByExample(PayInOrderDetailExample example);
+    long countByExample(PayInOrderDetailExample example);
 
     int deleteByExample(PayInOrderDetailExample example);
